[DATA-1165] Write unittest for name_suffix-function, move function to helm-utils
