+++ tests/utilities/test_helm_utilities.py
@@ -1,4 +1,5 @@
-from src.mpyl.utilities.helm import convert_to_helm_release_name, shorten_name
+from src.mpyl.utilities.helm import convert_to_helm_release_name, get_name_suffix
+from tests.test_resources import test_data
 
 
 class TestHelm:
@@ -9,5 +10,8 @@ class TestHelm:
         )
         assert convert_to_helm_release_name("my.project", "") == "my-project"
 
-    def test_shorten_name(self):
-        assert shorten_name("my-project") == "mp"
+    def test_get_name_suffix_for_pr_target(self):
+        assert get_name_suffix(test_data.RUN_PROPERTIES) == "-pr-1234"
+
+    def test_get_name_suffix_for_prod_target(self):
+        assert get_name_suffix(test_data.RUN_PROPERTIES_PROD) == ""
